requirements.txt
CTkMessagebox==2.5
CTkTable==1.1
customtkinter==5.2.2
darkdetect==0.8.0
et-xmlfile==1.1.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
